Safe components to own file
